Handle DO's paginated results in its API.
Without this, a large DO account will randomly break get_record(),
because get_record() only looked at the first page, before, and
then it would just create a new record from scratch, potentially
leaving you with hundreds of duplicate A records in your account
(which just makes the account even larger!).

json_value had to go because its awk program was too simple to handle
the recursiveness of json; in particular, : is a delimiter in json,
but it's also a delimiter in URLs, so getting the next page's URL
would break. I replaced it, and awk, with a new dependency on
https://stedolan.github.io/jq/, which is unfortunate but necessary.
There's no way to correctly parse javascript in a bash one-liner.